+++ lib/umami_client.rb
@@ -19,6 +19,7 @@ require_relative "umami_client/teams"
 require_relative "umami_client/links"
 require_relative "umami_client/pixels"
 require_relative "umami_client/admin"
+require_relative "umami_client/country_trust"
 require_relative "umami_client/rails"
 require_relative "umami_client/client"